upgrade dependencies

+++ modules/supabase/supabase.ts
@@ -3,37 +3,40 @@ import { AppState } from "react-native";
 import { storage } from "../storage/storage";
 // import "react-native-url-polyfill/auto";
 
-export const supabaseClient = createClient(
-  process.env.EXPO_PUBLIC_SUPABASE_URL || "",
-  process.env.EXPO_PUBLIC_SUPABASE_ANON_KEY || "",
-  {
-    auth: {
-      storage: {
-        getItem: async (key) => {
-          if (typeof window === "undefined") {
-            return null;
-          }
-          return storage.getString(key) ?? null;
-        },
-        setItem: async (key, value) => {
-          if (typeof window === "undefined") {
-            return;
-          }
-          storage.set(key, value);
-        },
-        removeItem: async (key) => {
-          if (typeof window === "undefined") {
-            return;
-          }
-          storage.delete(key);
-        },
+const supabaseUrl = process.env.EXPO_PUBLIC_SUPABASE_URL || "";
+const supabaseAnonKey = process.env.EXPO_PUBLIC_SUPABASE_ANON_KEY || "";
+
+if (!supabaseUrl || !supabaseAnonKey) {
+  throw new Error("Missing Supabase URL or Anon Key.");
+}
+
+export const supabaseClient = createClient(supabaseUrl, supabaseAnonKey, {
+  auth: {
+    storage: {
+      getItem: async (key) => {
+        if (typeof window === "undefined") {
+          return null;
+        }
+        return storage.getString(key) ?? null;
       },
-      autoRefreshToken: true,
-      persistSession: true,
-      detectSessionInUrl: false,
+      setItem: async (key, value) => {
+        if (typeof window === "undefined") {
+          return;
+        }
+        storage.set(key, value);
+      },
+      removeItem: async (key) => {
+        if (typeof window === "undefined") {
+          return;
+        }
+        storage.delete(key);
+      },
     },
-  }
-);
+    autoRefreshToken: true,
+    persistSession: true,
+    detectSessionInUrl: false,
+  },
+});
 
 // Tells Supabase Auth to continuously refresh the session automatically
 // if the app is in the foreground. When this is added, you will continue
